requirements.txt
furl>=2.1.4
loguru>=0.7.3
lz4>=4.4.3
numpy>=2.3.3
orjson>=3.10.15
pillow>=11.2.1
pyglm>=2.8.2
PySide6>=6.8.2.1
PyYAML>=6.0.2
QtAwesome>=1.4.0